Feat: #5 - QueryDsl 설정 및 PR 리뷰 반영
1. QueryDsl 설정 추가 및 테스트 코드 작성
2. PR 리뷰 내용 반영

+++ api-server/src/main/java/com/daangndaangn/apiserver/entity/user/User.java
@@ -31,7 +31,7 @@ public class User extends AuditingCreateUpdateEntity {
     private String profileUrl;
 
     @Column(nullable = false)
-    private long manner;
+    private float manner;
 
     @Builder
     private User(String email, String nickname, String location, String profileUrl) {
@@ -39,6 +39,6 @@ public class User extends AuditingCreateUpdateEntity {
         this.nickname = nickname;
         this.location = location;
         this.profileUrl = profileUrl;
-        this.manner = 0;
+        this.manner = 0.0f;
     }
 }
